requirements.txt
agent-framework-azure-ai==1.0.0b260107
agent-framework-core==1.0.0b260107
azure-ai-agentserver-core==1.0.0b7
azure-ai-agentserver-agentframework==1.0.0b7
azure-identity>=1.15.0
python-dotenv>=1.0.1
debugpy>=1.8.0
agent-dev-cli
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.41b0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation-asyncio>=0.41b0
